Split volofile to different parts
Now build/clean/appcache are different parts of the volofile, and
deploy calls all of them if necessary.
Change main.css to import datepicker css, so it is not necessary to
include the whole datepicker project in the final output anymore.
Change index and 404 to not include datepicker css.
Change build css to not keep lines.